refactor: 루트뷰 변경 구현 in appmain, locationmanager 이후에 항상  설정으로 이동"
